requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
gevent==24.2.1
greenlet==3.0.3
idna==3.7
NorenRestApiPy==0.0.22
numpy==1.26.4
pandas==2.2.2
psycopg2==2.9.9
pycparser==2.22
pyotp==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
setuptools==69.5.1
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
websocket==0.2.1
zope.event==5.0
zope.interface==6.4
opencensus-ext-azure==1.1.13